--- packages/config/lib/index.d.ts
@@ -1,1 +1,0 @@
-export declare const allowed: (obj = {}) => Record<string, any>
