+++ packages/data-context/src/sources/MigrationDataSource.ts
@@ -22,10 +22,6 @@ import {
   getIntegrationTestFilesGlobs,
 } from './migration'
 
-import { allowed } from '@packages/config/lib'
-import * as plugins from './migration/initOldPlugins'
-import * as configModule from '@packages/server/lib/config'
-
 import type { FilePart } from './migration/format'
 import Debug from 'debug'
 
@@ -325,28 +321,28 @@ export class MigrationDataSource {
     }
   }
 
-  async initializePlugins (cfg: OldCypressConfig): Promise<OldCypressConfig[]> {
-    return Promise.all((['e2e', 'component'] as const).map(async (type) => {
-      // only init plugins with the
-      // allowed config values to
-      // prevent tampering with the
-      // internals and breaking cypress
-      const allowedCfg = allowed(cfg)
+  // async initializePlugins (cfg: OldCypressConfig): Promise<OldCypressConfig[]> {
+  //   return Promise.all((['e2e', 'component'] as const).map(async (type) => {
+  //     // only init plugins with the
+  //     // allowed config values to
+  //     // prevent tampering with the
+  //     // internals and breaking cypress
+  //     const allowedCfg = allowed(cfg)
 
-      const modifiedCfg = await plugins.init(allowedCfg, {
-        projectRoot: this.ctx.lifecycleManager?.projectRoot,
-        configFile: path.join(this.ctx.lifecycleManager?.projectRoot, 'cypress.json'),
-        testingType: type,
-        onError: (err: Error) => {
-          this.ctx.coreData.baseError = err
-        },
-      })
+  //     const modifiedCfg = await plugins.init(allowedCfg, {
+  //       projectRoot: this.ctx.lifecycleManager?.projectRoot,
+  //       configFile: path.join(this.ctx.lifecycleManager?.projectRoot, 'cypress.json'),
+  //       testingType: type,
+  //       onError: (err: Error) => {
+  //         this.ctx.coreData.baseError = err
+  //       },
+  //     })
 
-      debug('plugin config yielded: %o', modifiedCfg)
+  //     debug('plugin config yielded: %o', modifiedCfg)
 
-      return configModule.updateWithPluginValues(cfg, modifiedCfg) as any as OldCypressConfig
-    }))
-  }
+  //     return configModule.updateWithPluginValues(cfg, modifiedCfg) as any as OldCypressConfig
+  //   }))
+  // }
 
   get step (): MIGRATION_STEP {
     return this._step
