+++ packages/server/lib/config.ts
@@ -3,7 +3,7 @@ import path from 'path'
 import Bluebird from 'bluebird'
 import deepDiff from 'return-deep-diff'
 import type { ResolvedFromConfig, ResolvedConfigurationOptionSource, AllModeOptions, FullConfig } from '@packages/types'
-import configUtils from '@packages/config'
+import * as configUtils from '@packages/config'
 
 import errors from './errors'
 import { fs } from './util/fs'
